Move tool form event handler to toolbox
